[Ashish] implement difficulty level

+++ src/main/java/org/models/CustomPlayer.java
@@ -15,6 +15,7 @@ public class CustomPlayer implements Player {
     public void setLastMove(int lastMove) {
         this.lastMove = lastMove;
     }
+
     public CustomPlayer(String name, Symbol symbol) {
         this.name = name;
         this.symbol = symbol;
@@ -30,7 +31,7 @@ public class CustomPlayer implements Player {
 
     @Override
     public void updateGrid(Board board, Scanner scanner) {
-        double maxCellNumber = Math.pow(board.getGrid().length, 2);
+        double maxCellNumber = (int) Math.pow(board.getGrid().length, 2);
         System.out.println(this + ", please choose as empty cell for your next move using the numbers 1 to " + maxCellNumber);
         int nextCell;
         try {
